+++ CHANGELOG.md
@@ -13,6 +13,9 @@
 - done command can close issue.
 - split commands by different entry points.
 
+### Changed
+- return the ability to set a parent for a task (for add command).
+
 
 ## [0.5.0] - 2022-12-05